bug: Scaling issue for crops
It appears that due to every panorama having the same dimensions in the Project Sidewalk database that the regression model works on this sv_image_y value assuming NO scaling factor. That means for larger GSV images, is sv_image_y is scaled * 1.23 this will result in a poorer crop of the object of interest. Poorer crop = more of surrounding environment/things included in crop which is not ideal.

To remedy this distance to object is calculated BEFORE scaling of sv_image_x and sv_image_y.