refactor: Additional docstring and unused line comments
Added comments to lines that are commented out in the code.
These are code blocks that were previously used for the XML metadata
for Google Street View. As this is no longer accessible, we now rely
on the input csv file which contains the metadata and do not need these
lines. They have been kept in for the Project Sidewalk team to decide
if they should be maintained or removed.